mediacapture-image/MediaStreamTrack-getConstraints.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/mediacapture-image/resources/imagecapture-helpers.js"></script>
<body>
<canvas id='canvas' width=10 height=10/>
</body>
<script>

const constraints = { whiteBalanceMode     : "single-shot",
                      exposureMode         : "manual",
                      focusMode            : "single-shot",

                      exposureCompensation : 133.77,
                      exposureTime         : 10000, // in nano-seconds.
                      colorTemperature     : 6000,
                      iso                  : 120.0,

                      brightness           : 3,
                      contrast             : 4,
                      saturation           : 5,
                      sharpness            : 6,
                      focusDistance        : 7,

                      // TODO: torch https://crbug.com/700607.
                    };

var canvas = document.getElementById('canvas');
var context = canvas.getContext("2d");
context.fillStyle = "red";
context.fillRect(0, 0, 10, 10);

// These tests verify that MediaStreamTrack.getConstraints() exists and that,
// returns the constraints passed beforehand with applyConstraints.
var makePromiseTest = function(c) {
  image_capture_test(async function(t) {
    var stream = canvas.captureStream();
    var videoTrack = stream.getVideoTracks()[0];

    // |videoTrack|'s capabilities gathering, just like the actual capture, is
    // a process kicked off right after creation, we introduce a small delay
    // to allow for those to be collected, since they are needed to understand
    // which constraints are supported in applyConstraints().
    // TODO(mcasas): this shouldn't be needed, https://crbug.com/711524.
    await new Promise(resolve => step_timeout(resolve, 100));

    var constraintsIn = { advanced : [ c ] };
    await videoTrack.applyConstraints(constraintsIn);
    assert_object_equals(videoTrack.getConstraints(), constraintsIn, "constraints");

    // Clear constraints by sending an empty constraint set.
    await videoTrack.applyConstraints({});
    assert_object_equals(videoTrack.getConstraints(), {}, "constraints");
  });
};

// Send each line of |constraints| in turn and then the whole dictionary.
for (key in constraints) {
  var one_constraint = {};
  one_constraint[key] = constraints[key];
  generate_tests(
      makePromiseTest,
      [[ 'MediaStreamTrack.getConstraints(), key: ' + key, one_constraint ]]);
}

generate_tests(makePromiseTest, [[
                 'MediaStreamTrack.getConstraints(), complete ', constraints
               ]]);

</script>
